+++ apps/client/src/features/editor/components/mention/mention-list.tsx
@@ -21,7 +21,8 @@ import {
 import clsx from "clsx";
 import classes from "./mention.module.css";
 import { CustomAvatar } from "@/components/ui/custom-avatar.tsx";
-import { IconFileDescription, IconPlus } from "@tabler/icons-react";
+import { IconPlus } from "@tabler/icons-react";
+import { PageIcon } from "@/components/ui/page-icon";
 import { useSpaceQuery } from "@/features/space/queries/space-query.ts";
 import { useParams } from "react-router-dom";
 import { v7 as uuid7 } from "uuid";
@@ -433,9 +434,7 @@ const MentionList = forwardRef<any, MentionListProps>((props, ref) => {
                     color="gray"
                     size="sm"
                   >
-                    {item.icon || (
-                      <IconFileDescription size={18} stroke={1.5} />
-                    )}
+                    <PageIcon value={item.icon} size={18} />
                   </ActionIcon>
 
                   <div style={{ flex: 1, minWidth: 0 }}>